+++ docs/source/CHANGELOG.rst
@@ -4,6 +4,8 @@ Changelog
 0.8.0
 -----
 
+(Unreleased)
+
 Bugfixes
 ~~~~~~~~